# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Pickup Sports App API
  version: 1.0.0
  description: |
    REST API for Pickup Sports App (Java 17, Spring Boot). Includes authentication, game management, search, notifications, profiles, venues, Nepal market features, media presign, and push subscriptions. JWT Bearer auth used for protected endpoints. Some features may be gated by feature flags or env properties.
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /auth/login:
    post:
      tags: [Authentication]
      summary: Login and obtain tokens (may return MFA challenge)
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Token pair or MFA requirement
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/TokenPairResponse'
                  - type: object
                    properties:
                      mfaRequired:
                        type: boolean
                      methods:
                        type: array
                        items:
                          type: string
                      challenge:
                        type: string
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthError'
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access and refresh tokens
      operationId: refresh
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: New token pair
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPairResponse'
  /auth/me:
    get:
      tags: [Authentication]
      summary: Current authenticated user info
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Principal info
          content:
            application/json:
              schema:
                type: object
                properties:
                  username:
                    type: string
                  roles:
                    type: array
                    items:
                      type: string
                  authenticated:
                    type: boolean
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AuthError:
      type: object
      properties:
        error: { type: string }
        message: { type: string }
        path: { type: string }
        status: { type: integer }
        timestamp: { type: integer, format: int64 }
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string }
    TokenPairResponse:
      type: object
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        refreshNonce: { type: string }
        tokenType: { type: string, example: Bearer }
        expiresInSeconds: { type: integer, format: int64 }
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken: { type: string }
        nonce: { type: string }
...